requirements.txt
sentry_sdk
python-dotenv